requirements.txt
fastapi
uvicorn[standard]
onnxruntime
numpy
pillow
requests
torch
torchvision
onnx
python-multipart